requirements.txt
asgiref==3.4.1
Django==3.2.8
pytz==2021.3
sqlparse==0.4.2
channels~=3.0.4
psycopg2~=2.9.1
channels_redis~=3.3.1
uuid~=1.30
djangorestframework~=3.12.4
drf_yasg~=1.20.0
django-cors-headers~=3.10.0